src/app/shared/components/spinner/spinner.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { SpinnerService } from 'src/app/core/services/spinner/spinner.service';

import { SpinnerComponent } from './spinner.component';

describe('SpinnerComponent', () => {
  let component: SpinnerComponent;
  let fixture: ComponentFixture<SpinnerComponent>;
  let service: SpinnerService;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [SpinnerComponent],
      providers: [SpinnerService]
    }).compileComponents();

    fixture = TestBed.createComponent(SpinnerComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should isShowSpinner be false', () => {
    expect(component.isShowSpinner).toBeFalsy();
  });

  it('should ngOnInit be called', () => {
    spyOn(component, 'ngOnInit');
    component.ngOnInit();
    expect(component.ngOnInit).toHaveBeenCalled();
  });
});
